ldd: pull out List function, rename to Paths
