+++ pkg/ldd/ldd_unix.go
@@ -5,21 +5,6 @@
 //go:build freebsd || linux
 // +build freebsd linux
 
-// ldd returns all the library dependencies of an executable.
-//
-// The way this is done on GNU-based systems is interesting. For each ELF, one
-// finds the .interp section. If there is no interpreter
-// there's not much to do.
-//
-// If there is an interpreter, we run it with the --list option and the file as
-// an argument. We need to parse the output.
-//
-// For all lines with =>  as the 2nd field, we take the 3rd field as a
-// dependency. The field may be a symlink.  Rather than stat the link and do
-// other such fooling around, we can do a readlink on it; if it fails, we just
-// need to add that file name; if it succeeds, we need to add that file name
-// and repeat with the next link in the chain. We can let the kernel do the
-// work of figuring what to do if and when we hit EMLINK.
 package ldd
 
 import (
@@ -105,11 +90,6 @@ func runinterp(interp, file string) ([]string, error) {
 		return nil, err
 	}
 	return parseinterp(string(o))
-}
-
-type FileInfo struct {
-	FullName string
-	os.FileInfo
 }
 
 func GetInterp(file string) (string, error) {
@@ -218,17 +198,4 @@ func Ldd(names ...string) ([]*FileInfo, error) {
 	}
 
 	return libs, nil
-}
-
-// List returns the dependency file paths of files in names.
-func List(names ...string) ([]string, error) {
-	var list []string
-	l, err := Ldd(names...)
-	if err != nil {
-		return nil, err
-	}
-	for i := range l {
-		list = append(list, l[i].FullName)
-	}
-	return list, nil
 }
